Replace Form and Window with InlineEditScreen
These components are confusingly named pieces of what once was
'editable-list-window'. This is a step towards renaming those
component styles and ensuring the component itself is used
consistently throughout the app.
