+++ src/components/Form/index.js
@@ -1,4 +1,3 @@
 export { default as Button } from './Button';
 export { default as ValidatedField } from './ValidatedField';
 export { default as MultiSelect } from './MultiSelect';
-export { default } from './Form';
